oci-ai-ds-model-server/mis_res_mgr/requirements.txt
fastapi[all]
requests
psycopg2-binary
